+++ tests/unit/str/test_str.py
@@ -6,13 +6,22 @@ from deduce.str import (
     RemoveValues,
     TakeLastToken,
     TitleCase,
+    UpperCase,
     UpperCaseFirstChar,
 )
 
 
 class TestStr:
-    def test_uppercase_first_char(self):
+    def test_uppercase(self):
+        processor = UpperCase()
 
+        assert processor.process("test") == "TEST"
+        assert processor.process("Test") == "TEST"
+        assert processor.process("TEST") == "TEST"
+        assert processor.process("123") == "123"
+        assert processor.process("test_123") == "TEST_123"
+
+    def test_uppercase_first_char(self):
         processor = UpperCaseFirstChar()
 
         assert processor.process("test") == "Test"