+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.3.0 (2023-10-25)
+
+### Added
+- lookup lists (and logic) for Dutch provinces, regions, municipalities and streets
+
+### Changed
+- name of `residences` annotator to `placenames`, now includes provinces, regions and municipalities
+- lookup lists (and logic) for residences
+- logic for streets, housenumber and housenumber letters
+
 ## 2.2.0 (2023-09-28)
 
 ### Changed